Memoize references and stop cloning the root
The github api has a _lot_ of references in it. This was taking longer
than my patience was willing to sit around for to run the test.

The first step was memoizing the dereferenced paths. This was the
simplest way to speed this up. It was still taking about 15 seconds on
my local machine to run the test. Which is still longer than I'm willing
to wait.

I looked a little further into this, and we were cloning the root json
node every time we were memoizing a reference because the jsonpath-rust
library trait that works with serde_json::Value requires ownership. The
implementation, however, doesn't. After looking at the latest commits,
there's a pretty simple way to use a borrowed serde_json::Value with the
crate, so I switched over to that. The tests now run in under a second
locally, and I'm going to leave at "good enough" for now.